many more fixes (now loads much faster as csrf is done in background)
