--- frontend/src/components/ProtectedRoute.tsx
@@ -1,44 +1,0 @@
-import React, { useEffect, useState } from 'react';
-import { Navigate, useLocation } from 'react-router-dom';
-import { Box, CircularProgress } from '@mui/material';
-import { verifyUser } from '../api/auth';
-
-interface ProtectedRouteProps {
-  children: React.ReactNode;
-}
-
-// Component to protect routes that require authentication
-export const ProtectedRoute: React.FC<ProtectedRouteProps> = ({ children }) => {
-  const [isAuthenticated, setIsAuthenticated] = useState<boolean | null>(null);
-  const location = useLocation();
-
-  useEffect(() => {
-    const checkAuth = async () => {
-      try {
-        const response = await verifyUser();
-        setIsAuthenticated(response.isValid);
-      } catch (error) {
-        console.error(error);
-        setIsAuthenticated(false);
-      }
-    };
-    checkAuth();
-  }, []);
-
-  if (isAuthenticated === null) {
-    return (
-      <Box display="flex" justifyContent="center" alignItems="center" height="100vh">
-        <CircularProgress />
-      </Box>
-    );
-  }
-
-  if (!isAuthenticated) {
-    // Save the attempted URL for redirecting after login
-    return <Navigate to="/login" state={{ from: location }} replace />;
-  }
-
-  return <>{children}</>;
-};
-
-export default ProtectedRoute;
